Changed the imports that follow from refactoring the utils
