+++ naslib/defaults/trainer.py
@@ -12,7 +12,7 @@ from fvcore.common.checkpoint import PeriodicCheckpointer
 
 from naslib.search_spaces.core.query_metrics import Metric
 
-from naslib.utils import utils
+from naslib import utils
 from naslib.utils.logging import log_every_n_seconds, log_first_n
 
 from typing import Callable
