+++ naslib/optimizers/discrete/bp/optimizer.py
@@ -9,7 +9,7 @@ from naslib.predictors.ensemble import Ensemble
 from naslib.search_spaces.core.query_metrics import Metric
 from naslib.search_spaces.core.graph import Graph
 
-from naslib.utils.utils import count_parameters_in_MB
+from naslib.utils import count_parameters_in_MB
 
 from fvcore.common.config import CfgNode
