+++ naslib/optimizers/discrete/re/optimizer.py
@@ -8,7 +8,7 @@ from naslib.optimizers.core.metaclasses import MetaOptimizer
 from naslib.search_spaces.core.query_metrics import Metric
 from naslib.search_spaces.core.graph import Graph
 
-from naslib.utils.utils import count_parameters_in_MB
+from naslib.utils import count_parameters_in_MB
 from naslib.utils.logging import log_every_n_seconds
 
 from fvcore.common.config import CfgNode
