+++ naslib/runners/nas/runner.py
@@ -21,8 +21,8 @@ from naslib.search_spaces import (
     TransBench101SearchSpaceMacro,
     NasBenchASRSearchSpace
 )
-
-from naslib.utils import utils, setup_logger, get_dataset_api
+from naslib import utils
+from naslib.utils import setup_logger, get_dataset_api
 
 from naslib.search_spaces.transbench101.loss import SoftmaxCrossEntropyWithLogits
