+++ naslib/runners/nas_predictors/runner.py
@@ -13,9 +13,8 @@ from naslib.search_spaces import (
     NasBench301SearchSpace,
     NasBenchNLPSearchSpace,
 )
-from naslib.utils import utils, setup_logger, get_dataset_api
-from naslib.utils.utils import get_project_root
-
+from naslib import utils
+from naslib.utils import setup_logger, get_dataset_api, get_project_root
 
 config = utils.get_config_from_args(config_type="nas_predictor")
