+++ naslib/runners/predictors/runner.py
@@ -39,8 +39,8 @@ from naslib.search_spaces import (
     TransBench101SearchSpaceMacro,
     NasBenchASRSearchSpace,
 )
-
-from naslib.utils import utils, setup_logger, get_dataset_api
+from naslib import utils
+from naslib.utils import setup_logger, get_dataset_api
 
 config = utils.get_config_from_args(config_type="predictor")
 utils.set_seed(config.seed)
