+++ naslib/runners/zc/runner.py
@@ -5,7 +5,8 @@ from  fvcore.common.config import CfgNode
 from naslib.evaluators.zc_evaluator import ZeroCostPredictorEvaluator
 from naslib.predictors import ZeroCost
 from naslib.search_spaces import get_search_space
-from naslib.utils import utils, setup_logger, get_dataset_api, get_zc_benchmark_api
+from naslib import utils
+from naslib.utils import setup_logger, get_dataset_api, get_zc_benchmark_api
 
 # Get the configs from naslib/configs/predictor_config.yaml and the command line arguments
 # The configs include the zero-cost method to use, the search space and dataset/task to use,
