+++ naslib/search_spaces/nasbenchnlp/graph.py
@@ -11,7 +11,7 @@ from naslib.search_spaces.core.query_metrics import Metric
 from naslib.search_spaces.core.graph import Graph
 from naslib.search_spaces.nasbenchnlp.conversions import convert_recipe_to_compact, \
 make_compact_mutable, convert_compact_to_recipe
-from naslib.utils.utils import get_project_root
+from naslib.utils import get_project_root
 from naslib.predictors.utils.encodings_nlp import encode_nlp
