+++ naslib/utils/get_dataset_api.py
@@ -2,8 +2,8 @@ import os
 import pickle
 import json
 
-from naslib.utils.utils import get_project_root
-from naslib.utils.utils_asr import from_folder
+from naslib.utils import get_project_root
+from naslib.utils.asr import from_folder
 
 """
 This file loads any dataset files or api's needed by the Trainer or PredictorEvaluator object.
